Split integration test file
we need to accommodate testing of a different flow where
requrests may be routed to another instance of the content
store, for drafts. this prepares us to do that.
